+++ CHANGES.rst
@@ -2241,6 +2241,8 @@ Other Changes and Additions
 
 - ``astropy.io.fits``
 
+  -  Fix use of quantize level parameter for ``CompImageHDU``. [#6029]
+
 - ``astropy.io.misc``
 
 - ``astropy.io.registry``